+++ part02/courseinfo/src/App.js
@@ -1,47 +1,29 @@
-const Header = ({ course }) => <h1>{course}</h1>;
-
-const Total = ({ sum }) => <p>Number of exercises {sum}</p>;
-
-const Part = ({ part }) => (
-  <p>
-    {part.name} {part.exercises}
-  </p>
-);
-
-const Content = ({ parts }) => (
-  <>
-    <Part part={parts[0]} />
-    <Part part={parts[1]} />
-    <Part part={parts[2]} />
-  </>
-);
+import Course from './components/Course';
 
 const App = () => {
-  const course = 'Half Stack application development';
-  const parts = [
-    {
-      name: 'Fundamentals of React',
-      exercises: 10,
-    },
-    {
-      name: 'Using props to pass data',
-      exercises: 7,
-    },
-    {
-      name: 'State of a component',
-      exercises: 14,
-    },
-  ];
+  const course = {
+    id: 1,
+    name: 'Half Stack application development',
+    parts: [
+      {
+        name: 'Fundamentals of React',
+        exercises: 10,
+        id: 1,
+      },
+      {
+        name: 'Using props to pass data',
+        exercises: 7,
+        id: 2,
+      },
+      {
+        name: 'State of a component',
+        exercises: 14,
+        id: 3,
+      },
+    ],
+  };
 
-  return (
-    <div>
-      <Header course={course} />
-      <Content parts={parts} />
-      <Total
-        sum={parts[0].exercises + parts[1].exercises + parts[2].exercises}
-      />
-    </div>
-  );
+  return <Course course={course}></Course>;
 };
 
 export default App;
