+++ part02/courseinfo/src/components/Course.js
@@ -1,0 +1,31 @@
+import React from 'react';
+
+const Header = ({ course }) => <h1>{course}</h1>;
+
+const Total = ({ sum }) => <p>Number of exercises {sum}</p>;
+
+const Part = ({ part }) => (
+  <p>
+    {part.name} {part.exercises}
+  </p>
+);
+
+const Content = ({ parts }) => (
+  <>
+    {parts.map((part) => (
+      <Part part={part} key={part.id}></Part>
+    ))}
+  </>
+);
+
+const Course = ({ course }) => {
+  return (
+    <div>
+      <Header course={course.name} />
+      <Content parts={course.parts} />
+      {/* <Total sum={parts[0].exercises + parts[1].exercises + parts[2].exercises} /> */}
+    </div>
+  );
+};
+
+export default Course;
